+++ collect_app/src/test/java/org/odk/collect/android/database/DatabaseInstancesRepositoryTest.kt
@@ -1,14 +1,10 @@
 package org.odk.collect.android.database
 
-import android.database.sqlite.SQLiteConstraintException
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import org.junit.Assert.assertThrows
-import org.junit.Test
 import org.junit.runner.RunWith
 import org.odk.collect.android.database.instances.DatabaseInstancesRepository
 import org.odk.collect.forms.instances.InstancesRepository
-import org.odk.collect.formstest.InstanceFixtures
 import org.odk.collect.formstest.InstancesRepositoryTest
 import org.odk.collect.shared.TempFiles.createTempDir
 import java.io.File
@@ -35,33 +31,5 @@ class DatabaseInstancesRepositoryTest : InstancesRepositoryTest() {
             instancesDir.absolutePath,
             clock
         )
-    }
-
-    @Test
-    fun save_failsWhenEditOfPointsAtItsOwnDbId() {
-        val instancesRepository = buildSubject()
-
-        val originalInstance = InstanceFixtures.instance(displayName = "Form1", instancesDir = instancesDir)
-        val originalInstanceDbId = instancesRepository.save(originalInstance)
-
-        val editedInstance = InstanceFixtures.instance(displayName = "Form1", instancesDir = instancesDir, editOf = originalInstanceDbId.dbId + 1)
-
-        assertThrows(SQLiteConstraintException::class.java) {
-            instancesRepository.save(editedInstance)
-        }
-    }
-
-    @Test
-    fun save_failsWhenEditOfPointsAtNonExistingDbId() {
-        val instancesRepository = buildSubject()
-
-        val originalInstance = InstanceFixtures.instance(displayName = "Form1", instancesDir = instancesDir)
-        val originalInstanceDbId = instancesRepository.save(originalInstance)
-
-        val editedInstance = InstanceFixtures.instance(displayName = "Form1", instancesDir = instancesDir, editOf = originalInstanceDbId.dbId + 100)
-
-        assertThrows(SQLiteConstraintException::class.java) {
-            instancesRepository.save(editedInstance)
-        }
     }
 }
