+++ forms/src/main/java/org/odk/collect/forms/instances/InstancesRepository.java
@@ -47,4 +47,6 @@ public interface InstancesRepository {
      * record in place.
      */
     void deleteWithLogging(Long id);
+
+    class IntegrityException extends RuntimeException {}
 }
